Add: defense programming to product and profile views

+++ products/views.py
@@ -1,10 +1,10 @@
 from django.shortcuts import render, redirect,reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import ProductFormAdmin
 
-# Create your views here.
 
 def all_products(request):
     """ Displays all products, including sorting and searching """
@@ -49,8 +49,13 @@ def product_description(request, product_id):
     return render(request, 'products/product_description.html', context)
 
 
+@login_required
 def add_product(request):
     """Add product to store through the admin"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductFormAdmin(request.POST, request.FILES)
         if form.is_valid():
@@ -70,8 +75,12 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store from the admin """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -95,8 +104,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfully deleted!')
